+++ src/giit/push_config.py
@@ -58,10 +58,13 @@ class PushConfig(object):
             config['exclude_patterns'] = []
 
         if not 'scope' in config:
-            config['scope'] = ['source_branch']
+            config['scope'] = ['branch']
 
         if not 'tag_semver_filter' in config:
             config['tag_semver_filter'] = None
+
+        if not 'branches' in config:
+            config['branches'] = []
 
         if not 'nojekyll' in config:
             config['nojekyll'] = False